Basic code to return rows from SELECT executions
This doesn't work yet because there appears to be a bug with the
implementation of `QueryResultWriter`. However, the shim-level code is
in place and should be correct. Note that it currently forcibly returns
all columns as strings and with empty flags; we'll need to expose more
information from Soup to do this properly.