Rudimentary support for AUTO_INCREMENT columns
This correctly sets the value for AUTO_INCREMENT columns without an
explicit value set on processing an INSERT. The AUTO_INCREMENT counter
is maintained in the shim, so it resets to zero on every restart.
Moreover, we don't handle duplicate IDs correctly, since the Soup
backend is happy to accept duplicates.