Fix returned row data; Django works!
Turns out the code previously returned all rows as strings, which
greatly confused the far end in two ways:
 1) Some row values were enclosed in superfluous quotes.
 2) TinyText records from Soup turned into strings with trailing,
    invisible null characters padding up to 15 bytes (the maximum
    TinyText length).

Getting the dance of converting Soup's `DataType` into MySQL types via
`ToMysqlValue` right is a bit tricky, and the current solution is a hack
in the absence of an `impl ToMysqlValue for DataType`, but it works well
enough to complete the Django setup process. The shim now crashes when a
new client connects via the web browser because the session state
associated with a single Soup backend connection doesn't get shared.
That's a solvable problem, though -- and getting here is progress!