Fix ExactMatch reuse bug
Building on the prior commits, this fixes the incorrect `ExactMatch`
reuse that we suffered from due to assuming the wrong semantics for QG
signatures.

The fix is a brute-force one because it uses the `exact_hash` of the QG
to decide equivalence: as per c16434c, this will only match queries that
have the same predicate order, even though different predicate orders
should also allow for `ExactMatch` reuse. Fixing this in the presence of
nested predicates is not easy: we'd need a data structure that *both*
allows order-independent comparison of predicates, *and* preserves the
necessary nesting hierarchy.