Avoid spurious ReaderOntoExisting reuse
Now that `ExactMatch` relies on matching exact query graphs, the
next case we check (adding a reader onto an existing query leaf) also
needs fixing. It is only legal to do this reuse if all predicates in the
incoming query have equivalents in the existing one (although parameters
can differ). The check for predicate equality is somewhat crude: as in
554bd25, it does not handle reordering under nested predicates
correctly.