Escape table and column names if they are keywords
Fixes an issue with the implementations of `Display` for various nom-sql
structures: queries with columns or tables whose names are also SQL
keywords would not correctly escape these names when displayed. This
caused "reflection" of queries through parsing and `format!()` to create
queries with invalid syntax and impossible to parse a second time.