Correctly handle primary keys defined inline
SQL allows two syntaxes for defining primary keys in CREATE TABLE:
 1) at the end, as PRIMARY KEY(column_name); or
 2) inline, next to the column definition (e.g., "col int PRIMARY KEY")

We parse both of them correctly, but previously only added a primary key
in Soup for the first syntax; the second (a `ColumnConstraint`) was
silently ignored. This adds a query processing pass that combines all
primary key definitions in `CreateTableStatement::keys`, though leaving
the column constraint in place also.