Gracefully handle failed heartbeats
When a worker fails to send a heartbeat to the controller, this can have
several reasons, some of which indicate disconnection. In
high-concurrency environments (e.g., on Travis), this can get triggered
spuriously at the end of tests when the controller has already exited,
but the worker tries to send one more heartbeat.

We should ultimately handle this better, but since heartbeats are only
used for the controller to detect worker failure at the moment, it's
okay to just ignore failed heartbeat sends on the worker side.