Fix bug: don't try to evict more bytes than exist
The state size we read in the worker is often stale, since the
`UpdateStateSize` packet can be delayed. It may hence fail to reflect
either (a) evictions that have since happened, or (b) "normal" record
deletions that have since happened. Unfortunately, since we use this
weakly-consistent value to decide how many bytes to evict, and since
we spin until we've evicted sufficiently many bytes, we can now get
stuck if we ask a domain to evict more bytes than it actually stores.
Instead, give up early if the domain's state is now empty.