Eviction mechanics for reader nodes
This puts in place the necessary APIs for eviction from reader state,
i.e., from an `evmap` via `dataflow::backlog`. The eviction works, but
the state size tracking is off: we cannot currently tell how many bytes
we actually ended up evicting, because the (random) key isn't chosen
until the eviction gets to `evmap`, and the random index we choose
cannot currently be used to look up the key ahead of eviction.

Will need to add either a `get_at_index` API to rahashmap, or figure out
a different way of picking keys to evict.