Start prepared statement support
So far, the applications we've been using did not use MySQL prepared
statements. However, the "trawler" benchmarker for Lobste.rs does. This
change puts in place the basic mechanics for "prepare": we extract the
query's schema, add the query to Soup, and respond to the client with a
statement ID and schema meta-data.

Still lacks support for actually executing prepared statements, and for
returning the correct parameter columns, but does already add the first
Lobste.rs query correctly.