Extract parameter columns from prepared statement
`on_prepare` must return a statement ID, the column schema of query
responses, and the set of parameter columns that an execution of the
prepared statement must specify. This implements the extraction of
parameter columns from the query, so that we return the correct data
structure.

We now get a bit further in handling a prepared statement from Lobsters.
Still need to ensure we return the correct column schema (specifically,
only the query-specified columns, not the parameter columns, which Soup
also returns). Also currently crashes with a weird binary protocol error
in `msql_srv` code...