Parse SEND_LONG_DATA commands
MySQL sends "long" parameters using `SEND_LONG_DATA` commands:

https://github.com/MariaDB/server/blob/10.3/sql/sql_prepare.cc#L73-L84

This happens, for example, if a `VARCHAR` field is used as a parameter.

This change adds support for parsing these commands, but leaves their
handling `unimplemented!()` as we'll need to decide whether to keep
parameter state in msql_str or force the implementing library to store
it (via `on_send_long_data` or similar).