Expand "table.*" to full schema
It's a bit unfortunate that we have to do this in the MySQL shim, since
Soup already performs the same expansion internally. It is required here
because the MySQL protocol expects the schema of the columns returned to
be specified alongside the data.

Lobsters user creation and tag creation work; we get halfway through
story creation before tripping over "IS NULL" (fix coming).