Cache ad-hoc queries to avoid duplication
This adds a local "query cache" that checks, for ad-hoc SELECT queries,
if we've seen this query before. If so, the cache supplies the Soup-side
identifier for the query, rather than triggering a migration on Soup.
(The migration actually does nothing: Soup detects that the query is
identical and simply installs a recipe-level alias.)

This improves performance, and reduces the number of getters we request
from Soup (and hence, the number of active TCP connections).

Lobsters story creation now succeeds for the full dataset, and much more
quickly than before.