Have global and thread-local query caches
When running with many client threads, we still triggered
unnecessary migrations on Soup, since the threads only cached prepared
statement and query identifiers locally. This also adds a global cache
that the threads consult if they miss in the thread-local cache.

Now only does a single migration when we see a query for the first time.