Differentiate between key for lookup and src cols
if we do state.lookup(&[i], KeyType::from(&r[j])), i and j are not
necessarily equal. We need to record *both* in a miss. i has to be used
when selecting what index to replay *into*, while j has to be used when
figuring out *which* key to replay based on the record that missed.