Support for prepared UPDATE statements
We can now prepare and execute any UPDATE statement that has parameters
in the WHERE clause (not in the SET clause, so far, though we can easily
support it if needed).

Also restructures the UPDATE code to share most of the logic between
prepared and ad-hoc queries. Tests still pass.