Handle schemas for columns without a table set
Computed columns are not associated with a table, but we still need to
return a MySQL column schema for them. Currently, we return an integer
column with no flags set, which is probably right for most aggregations.
However, it doesn't seem like client libraries really check this schema,
so we should be fine either way.