Fix TopK: needs bogokey if there are no parameters
Queries with a TopK operator would not actually work correctly when
they had no parameters. We use the parameters as group columns for the
TopK, but in the absence of any parameters (e.g., for the "frontpage"
query in Lobsters), we -- lacking a group column -- would simply pick
the first column of the view to group by (here, "stories.id"). This
returns incorrect results because it now makes every story a group!

To fix it, we now introduce the bogokey (used for lookups into views
without parameters) earlier in the graph, and group the TopK operator by
this column. There are some potential problematic downstream effects
(particularly with security policies and UNIONed compound selection),
but those don't occur in Lobsters as I've tried to avoid foreseeable
pitfalls.

Also required some unfortunate cloning because NLL cannot figure out the
lifetimes. I suspect splitting up the giant method that all of this
action happens in would help the compiler.