Add a cache of pre-parsed queries
Re-parsing queries every time we see them is expensive. It's cheaper to
hash the query to see if we already have a parsed version of it, and to
only do the parsing if we don't have one.

This affects both `on_query` and `on_prepare`; particularly with the
latter, it results in a significant reduction in CPU load when running
the Lobsters workload. This may suggest that we re-prepare queries
rather more than we think we do (even though no migrations happen on
Soup as a result, and our local mutator/getter cache short-circuits
other RPCs).