Bug: don't try to evict too much from reader
The state sizes that the worker bases its eviction decisions on can be
somewhat stale as they are only updated occasionally. However, we
previously took the desired number of bytes to evict as gospel, without
checking it against the current state size when processing the
`Evict` message at the domain. This could lead to use trying to evict
more state as there actually is (particularly if we sample the state
size in between issuing an eviction and it being reflected in the state
size).