Fast-path for negative for unique single state
This change may be a little controversial. We assume (in release mode)
that if we're asked to remove a row, and there is only one row in the
current bucket, then the two are the same. This matches the grander
assumption in Soup that negatives always follow a positive, but may be
violated (now silently) in buggy cases.