Track global predicates in query graph
The short-cuts we took in the early days of Soup in terms of query
support keep coming back to bite us. We decided to skip support for
global predicates because they're harder to deal with, but OR
expressions are global predicates unless both sides refer to only one
and the same table. Moreover, predicates over computed columns (like
filters over arithmetic projections) are also global predicates
according to the Finkelstein algorithm.

This change makes the query graph track the global predicates correctly.
We still need to add code to actually insert filters based on global
predicates.