Revert part of 8f55e5a for compound SELECT unions
It turns out that 8f55e5a was too zealous in checking column equality in
the special case of a UNION over the nested queries of a compound
SELECT. Since we add these queries as views in their own right, we
rewrite the table name on their leaf node's columns to the view name,
but the union expected matching full column specification on both sides.
This change reverts that to matching by name, but we should fix this in
the future as it currently will be incorrect if two columns of the same
name appear on either side of the union.