Fix alias removal for table aliases
This drops aliases from `Table` structures after rewriting any columns
that refer to the table by its alias. This is required so that `Table`
structures always correctly compare equal in the query planning code.

In the long term, we'll want to get rid of this rewrite, since it
precludes doing self-joins in the natural way.

All tests pass again.