Consider global predicates for pre-agg filters
We push predicates that act on aggregated columns above the aggregations
in a query. However, we previously only considered local predicates
(since Soup didn't use to support global ones) for this lifting; this
change adds partial support for lifting of global predicates.

We cannot yet lift predicates that operate on the output of a literal or
arithmetic projection.