Use less beefy client
Still saturates the server workers, and is more likely to be available.

Note that the cload early termination limit has also been removed.
That's because the client load sometimes spikes severely right after
warmup finishes (probably because of lots of threads exiting at once),
which artificially inflates the load avg by a lot.