Port shim to use new API subcrate
This ports the MySQL shim to work with the latest master revisions of
Soup, which include a separate "api" subcrate that implements only the
necessary wire protocols for communication between a client application
and the Soup controller and souplet instances.

Still depends on RocksDB (via distributary::basics), but that
dependency is going to go away shortly.

The change much reduces the number of dependencies for the shim; speeds
up build times; and cleans up some interfaces.

Note that since the "api" subcrate relies on the 'failure' crate to
indicate RPC errors, we now have to do some awkward conversions from
failure::Error to errors that io::Result (which msql_srv expects) can
represent.