Attempt test harness for ControllerInner
This is an attempt at writing a test harness that allows more systematic
testing of `ControllerInner` with mock workers and without the
complexity of the multi-layer controller API. This seems like a good
plan in theory, but I ran into lots of obstacles, largely related to
the mock workers actually needing to respond to messages correctly, the
domains having to be reachable and processed by workers, etc.

Checkpointing this for now, but I may revert it. Currently, the test
gets stuck; I fear the mock worker will need to be almost as
sophisticated as the real thing in order for tests to pass.