Revert "Attempt test harness for ControllerInner"
This reverts commit 6f44eab915c48ffc2d6f44a3137e0f620f817438
and 393f22ae23b8f7b03a567e7a3ba4a90f566eee4b, which tried to add a test
harness that turned out to be too complex (for now, at least). I'm
leaving the commits in the history for merging in case we want to
resurrect the `MockWorker` test framework for `ControllerInner` in the
future.