Add a reuse test covering query prefix extension
We already have tests for the ExactMatch and ReaderOntoExisting
reuse cases, but no direct test for ExtendExisting, i.e., the case where
a new query shares the prefix of an existing one. This test exercises
this case; I wrote it while debugging the broken security universe
creation code (which uses a slight variant of the reuse tested here).