Rely mostly on Noria-provided schemas
This changes (again) how the shim handles schemas. In particular:

 * it now uses the schemas carried by `View` and `Table` objects
   obtained from Noria directly whenever possible;
 * it avoids using the global `table_schemas` hashmap, which reduces
   thread synchronization; and
 * it relies on Noria to expand stars in queries, rather than doing it
   locally.

Passes most tests, but some issues remain; in particular, Noria
currently specifies TEXT as the type of all columns in views.

Another issue is that parameter columns of SELECT and UPDATE queries
can come from *any* table or view mentioned in the query, and need
not be present in the output (e.g., "WHERE uid = ?" with uid not
projected in the output view). This isn't a huge issue at the moment
since Noria's views always contain and expose the parameter columns even
if the query does not explicitly project them, but handling it correctly
in all cases run requires the shim to be able to query Noria for the
schema of any column in any view (including internal ones).