Add (failing) test for query over aliased view
This test exercises a bug found while developing the stateless shim: due
to recipe-level aliasing of identical queries, some views may not exist
at the lower levels of the SQL processing code. Writing queries that
define other views that refer to these aliased views triggers a crash,
since the SQL layer has never heard of the aliased view.

The fix for this is probably to remove recipe-level deduplication of
identical queries and rely on the `ExactMatch` reuse detection in our
SQL layer. This may hurt performance a bit, for two reasons: (1) it will
do a migration even for identical queries, though this migration will
add nothing; (2) it will duplicate even identical queries when reuse is
turned off.