Specifying field names on INSERT is optional
Removes an old hack: when an `INSERT` query specified no field name
list, we would name the fields 0, 1, etc. in the absence of schema
information. This change makes the field name list an `Option`, so that
consumers of nom-sql handle this themselves (and, in particular, can
decide whether to obtain schema information).