Reset aliases when sanitizing leaf columns
Fixes an insidious column aliasing bug. Since we weren't clearing the
aliases of columns on leaf views, it was possible to refer to columns
with their "internal" names inside a query even from other queries. In
particular, two aggregations that happen to have the same column name
("votes") in two distinct intermediate views could look identical,
thanks to aliasing and the fact that we do not compare the `function`
field when checking columns for aliasing/equality (only table and name
matter). This led to a join deciding to only output one of these
aggregation columns, causing all kinds of confusion and breaking
Lobsters.

The fix clears any existing aliases at the view at the bottom of a
query. I *think* this is always correct, since we never rely on
preserving a column's identity *through* a view definition.