Name queries with hashes, not sequential counter
This ensures that query names are deterministic, and thus avoids bugs
when restarting the shim or when running more than one shim.

Previously, "q_1" prior to and after a restart, or in different shim
instances, could refer to completely different queries, but Noria would
quietly reuse the existing query by name and return incorrect results.
Whoopsy.