Fix create_view test, add prepared_select test
It turns out that MySQL uses the text protocol for ad-hoc queries and
the binary protocol for prepared statements (see
https://mariadb.com/kb/en/library/2-text-protocol/,
https://mariadb.com/kb/en/library/3-binary-protocol-prepared-statements/).

The MySQL client library therefore exposes results from ad-hoc queries
as `Bytes` structures containing strings, even if the shim writes values
with different, schema-accurate types. When using prepared statements,
the return values have the correct types specified in the schema. This
is pretty counter-intuitive, but we can't do much about it.

Adapt tests to expect query results in appropriate types, and document
this issue.