Make block_on resilient to having to try again
Previously, the first call to `poll_fn` would immediately unwrap the
`Option`. While this is _normally_ fine, it causes issues when
`blocking` returns `NotReady`, since `poll_fn` will the be called again,
but this time with `wrap` set to `None`. To fix this, we make sure to
only `take` the function when blocking succeeds (i.e., when the passed
closure gets executed).