Avoid linear scan of tags after replays
Previously, we would keep around the `HashSet` for any tag we had
previously buffered replays for to save on later memory allocation
costs. That was all well and good, but we did it by keeping the entry in
buffered_replay_requests, which is indexed by tag. To see if the timer
had expired for any entries, we had to walk that list, which means that
for a domain with many tags, it has to basically walk all of its tags
_for every packet_. Ouch. We now give up that small memory allocation
"optimization" for the much better scalability of, well, not doing it.