Avoid crash on write-replay conflict
Previously, a union that receives a replay from one ancestor, then an
update from a different ancestor that it has never seen a replay from
before, would crash. This happened because it was trying to find the
replay column for that ancestor. However, in this case it knows that it
cannot have any buffered replay pieces from that ancestor for _any_ key,
and therefore the write should just be forwarded.