Upgrade nom-sql; fixes Lobsters query misparse
This moves Noria onto using the latest (v0.0.6) nom-sql parser, which
includes a slew of fixes and improvements. Most importantly, I fixed a
bug that confused "FULLTEXT INDEX ..." for a column called "fulltext"
followed by an index definition, because the parser did not know that
FULLTEXT is a SQL keyword. As a result, several Lobsters tables had an
extra phantom "fulltext" column, and queries over these tables referred
to that column.

Also upgrades the recipe parsing logic to nom v4.