Fix comparison signed/unsigned integer comparison
The "_ => false" case in the implementation of PartialEq for
DataType led to #121 neglecting to correctly implement comparisons
between signed and unsigned integer types.

This adds the right comparisons, plus test to check that they work.
Tests for the Ord trait and the new types are still missing.

(Relates to https://github.com/mit-pdos/noria-mysql/pull/9)