Store only one operator on a reader
Backported from upstream at 463e0726. Original message:

get the query operator via MirNodeType

Note from committer: I think what's going on here is that it turns out
every reader's job is only to serve *one* kind of query, so we only need
to store one operator on the reader. Upstream also had this as an
Option<Operator>, but I deeemed that unnecessary as we can just default
to Equal.