Don't accidentally send acks on outdated streams
Previously, the following could happen:

 - Stream 1 connects; is issued stream token A
 - Stream 1 issues a write W
 - Stream 1 disconnects (probably a client error)
   At this point, A is removed and considered re-usable
 - Stream 2 connects; is also issued stream token A
 - W completes; ack is sent on A, which is stream 2

This is obviously a problem. This manifested itself in a slightly
different way where W completes before A is re-used, which caused a
"stream does not exist" crash, but the underlying cause was the same.

This patch adds an epoch to each stream token, and ignores acks for
tokens whose streams have gone away.