Integrate an interval tree in evbtree
Add an interval tree field to evbtree::Inner, to track the intervals of
keys that we have received (rather than just the keys themselves) and
expose its properties in the API of the evbtree. Concretely, this means
that the `range` method on ReadRef now returns Result<RangeIter, Miss>,
where Miss is a struct with a list of intervals that aren't present in
the requested range. This is also exposed all the way up through the
backlog.

Backported from 136f005634fb13d10e83ff1f8edce4ee6b3b5ae9