Fix view resolution with aliases
When a recipe contains multiple identical queries, we alias them onto a
single materialized view. When trying to create a read handle using the
name of the aliased view, the RPC handler to return a `ViewBuilder` was
correctly resolving the node ID for the final query node, but then
sought to find a `Reader` node with the alias name, which of course does
not exist. The solution is to use `Recipe::resolve_alias` to translate
the name prior to the BFS in `controller::inner::find_view_for`.