step 1: buffer replays separately per shard
The work for this patch was made in
https://www.youtube.com/watch?v=kiMUI0y91YI.

This patch also adds a test case for the known-bad case, which is not
yet fixed by the patch (it hangs).

This patch also leaves some work for later, namely that the union
buffers should also be per _tag_. This gets complicated, because you
won't know what key to look up until you look at the tag, whereas the
union code currently assumes that the key column is always the same.
Should be relatively straightforward to fix though.