Make last sharder route replay correctly
A sharder should _generally_ send replays to all shards. The exception
to this is the sharder is the last sharder before a sharded target. In
that case, the replay should only be aimed at the shard that requested
the replay, not any others (as they'd be confused at the arrival of a
replay they never asked for).