Force vote clients to complete all work
Without this, a vote client could just keep piling on work into the
"last" batch, which is then never issued, and therefore never measured.
This resulted in inaccurate throughput and latency measurements. It was
especially true for read-heavy benchmarks, where the number of writes is
small enough that you might not notice the drop in overall throughput,
but the surge in write latency won't be visible.