vote: saner batching strategy
Previously, vote would batch requests for _at least_ 1ms, and with no
upper limit. This was terrible. It meant that when the system _was_
keeping up, users would see a latency >500µs, and when it was _not_
(even if briefly), batches would grow huge, and requests enqueued a long
time ago still had to wait for the most recently added to the batch to
complete.

Now, vote batches for _at most_ 1ms, and issues batches pretty much
whenever the backend allows it to. Let's see how that turns out.