Implement all remaining column types
Implement schema::convert_column for all column types in the
nom_sql::SqlType enum. A lot of this was based on guesswork on my part,
as the mysql binary protocol documentation (I was mostly looking at
https://dev.mysql.com/doc/internals/en/binary-protocol-value.html) is a
little thin, so more testing is probably in order to make sure all of
this works. That said, with this change the following query on the
employees.tiny.noria.sql test db is working:

    SELECT first_name, employees.emp_no, dept_no
    FROM employees
    JOIN dept_emp ON employees.emp_no = dept_emp.emp_no