Make the table expression (FROM ...) optional
`SELECT 1` is valid SQL - per the ANSI standard, the *entire* FROM
expression is optional, and you can  just select a list of columns
directly. This is the most minimally-impacting change to implement
that - the BNF grammar in the SQL standard has the table_expression as
its own AST node which is optional as a whole, but I've avoided that
here since it's a pretty big breaking change.