Update to the latest version of nom-sql
Update to the latest (forked) version of nom-sql, and update the API to
be compatible across all calls. This is pretty large, but it's mostly
mechanical renames, with the most notable exception being the change in
the arithmetic conversion logic - since nom-sql supports nested
arithmetic now, we have to as well.