Wrap rustc in sccache
Add a pre-built sccache binary to the docker image we're building in,
and write a cargo config to root's home directory to tell cargo to use
sccache as a rustc wrapper. This is probably not *enough* to get fully
cached compilation quite yet - the next step is probably to tell sccache
to use s3 as a backing store.