Disable incremental compilation in CI
Per sccache's docs, any compilation requests with incremental
compilation enabled are uncacheable. Based on the hypothesis that the
majority of builds will be either changing the minority of files, or
running on an agent that doesn't happen to have run builds before (due
to autoscaling), this will *ideally* speed up enough builds to make
it worth it.