Handle BETWEEN conditions
Add explicit handling for BETWEEN conditions in the SQL engine. Since
BETWEEN is really just syntactic sugar, this is implemented as a rewrite
pass into the conjunction of a pair of comparisons. Concretely, the
following query:

    SELECT * FROM table WHERE x BETWEEN 1 AND 2;

Gets rewritten to:

    SELECT * FROM table WHERE x >= 1 AND x <= 2;

This doesn't work with negation (x NOT BETWEEN or NOT (x BETWEEN))
because the algorithm used to convert SQL queries into dataflow doesn't
currently support top-level disjunction - this is reported at
https://app.clubhouse.io/readysettech/story/75 will have to be explored
later