Vendor a clone of evmap
Add a clone of the source code of EVMap, in preparation to update it to
be backed by a BTreeMap instead of a HashMap. The idea here is that this
can be used to support things like ordering, looking up ranges (<=, >=,
etc), and potentially even LIKE prefix lookups.

This approach (cloning all the source code of evmap and swapping out the
innermost type) is unfortunate, but I couldn't think of a better
solution that satisfied our current needs.