Consider pagination part of a query graph's hash; remove hack
Add pagination (comprised of order, optionally limit, and optionally
offset) to the QueryGraph type, and use that when computing the query's
hash. This prevents the query deduplication mechanism from
considering:

    SELECT * FROM articles;

to be the same query as:

    SELECT * FROM articles ORDER BY column LIMIT 5;

Before this change, issuing those two queries in order would panic on
looking up the view for the second one, as it would be considered
"duplicate" even though it didn't exist verbatim in the graph.

Additionally, we remove eta's hack for non-TopK ORDER BY, since the
root cause is now fixed (we think).