Wrap the result of Node::process in a struct
Wrap the 3-tuple of (misses, lookups, captured) returned by
Node::process in a new NodeProcessingResult struct. This is distinct
from ops::process::ProcessingResult (which perhaps should be renamed now
to IngredientProcessingResult or something) because it tracks the final
captured results of the ingredient rather than only the records
forwarded on by processing - the distinction of which I am
not *entirely* clear on yet, but which I think has something to do with
materialization.

In addition to just being good style, this also prepares for adding an
additional field to this struct (and ProcessingResult) to track the case
where we need to not just mark that we missed on a key, but *fully*
evict all missed keys all the way down the chain, which is currently
special-cased to only the case where joins miss on something that exists
downstream.