Don't make builds depend on each other
The original idea here was that if eg nom-sql failed to build, then
noria and noria-mysql would *definitely* fail to build - so we should
build the roots of the intra-repo dependency graph first. This hurts the
successful case, though, since sharing of build artifacts between build
steps is less than perfect - so let's try this way, with the
understanding that this'll slow things down if builds do actually fail.