Permissive query parsing flag
Add a flag to allow for queries to fail to parse but return a success to the
client, along with a warning when that behavior happens. This is useful for
frontends like python that sometimes issue spurious queries that we can't parse:

For example,

```
import mysql.connector
conn = mysql.connector.connect(host='127.0.0.1', user='foo', password='bar')
```

induces the following error in noria-mysql:

```
noria_mysql::backend: query can't be parsed: ""SET NAMES \'utf8mb4\' COLLATE
\'utf8mb4_general_ci\'"" query=SET NAMES 'utf8mb4' COLLATE 'utf8mb4_general_ci
```

which in turn (with readysettech/noria-mysql#3) causes an exception in Python,
when the client would otherwise work fine.