Split out all logic pertaining to noria from the main backend.
Added a trait for a reader and writer and split the noria connector
into the reader and writer.

We need to distinguish between a reader and writer in order
to support writing to an underlying mysql implementation.
Furthermore, there are obvious use cases for reading and writing
to multiple sorces to check for parity.  I did not change any
of the body of these functions other than to remove dead code
in previous commits unless it was totally necessary to make
the refactor functional.