Maintain state inside a fully materialized topk
If a topk node ever drops to fewer than k records (due to negatives) we
currently have no way of issuing a query to bring us back up to k
records - previously, we would just panic if this ever happened. The
proper way to handle this depends on whether we are partial or full:

- If we are partial, we want to do a similar thing to what joins do
  where we issue a full eviction of all state all the way down all
  replay paths that go through the node, so that the next read triggers
  an upquery and full replay
- If we are fully materialized, we *can't* evict, so we need to maintain
  an internal cache of *all* rows - not just our top k rows - that we
  can then read from if we ever drop below k records.

Since due to a bug in MIR topk nodes are *always* fully materialized[1],
this commit implements the second case, leaving the first case
unimplemented.

[1]: MIR is putting a bogokey-projection node *after* every topk in
     addition to before, which means that the reader can't see the
     bogokey through the topk, meaning the reader must be full hence the
     topk must be full.