Don't re-project a bogokey on top of topk
When we decide to add a topk node to a query that doesn't have a
group_by, we have to project out a literal bogokey so that the topk
has *something* to group by / key off of. We then *also* always project
out the list of columns at the end of every query, so that we are
guaranteed to have the expected results from the query. Previously, this
would mean that we would *reproject* a literal bogokey out of the topk,
even though the columns emitted from the topk included the previously
projected literal bogokey. This commit changes that to track whether we
have already projected out a bogokey column, and if so just use that
when projecting out the final results. This allows the final reader node
at the very least to remain partially materialized, even though the topk
itself still has to be full because it's below a bogokey projection.