Differentiate between binary and unary operators
The Operator enum had a Not member for some reason - for one, this was
never actually constructed, but for another this doesn't make
sense in all the places where we use operators, all of which have an
explicit left-hand and right-hand side. This commit renames Operator to
BinaryOperator for clarity, and removes Not from the enum.