Add evbtree to CI
Add steps to CI to build and test evbtree, check its formatting, and run
Clippy on it. The latter also includes some automatically generated
fixes for clippy lints that were firing on evbtree.

Noria depends on evbtree in CI as of this commit even though it doesn't
actually use it yet, because it's going to very soon.