Don't gate eviction behind the eviction feature
The eviction feature for evmap previously required the use of the
indexmap data type as the underlying data type so that it could evict
indices - this was of course not compatible with the reimplementation to
use a btree as the backing storage instead. This commit changes
empty_random to instead just clone keys into a sequence to evict, which
while slightly slower than indices does still work with btrees for now,
and perhaps in the future we can optimize it further.