Split up readers::handle_message
Split up the handle_message function in noria_server::worker::readers
into separate functions for each of the types of message we might
receive, which are represented by members of the ReadQuery enum. This
makes things a little easier to reason about locally in my opinion,
especially as we extend these functions to get a little bit more
complex.