Add a KeyComparison enum for top-level queries
Define a new KeyComparison enum for top-level queries. This includes
Equal, which provides the current query semantics and is the default
version returned by Into, but also provides LessThan, LessThanOrEqualTo,
GreaterThan, and GreaterThanOrEqualTo. There's an implementation of
RangeBounds, too, which will allow us to pass it directly to functions
like `range` on an evbtree.

The keys that the KeyComparison wrap are instances of Vec1, a new crate
I've added as a dependency, which guarantees that it will never be
empty. I've done this because performing a lookup on an empty key is
meaningless - and in fact it used to just fail an assert and panic
before in multiple places! - and this lets us use the type system to
guarantee this instead. Yay, types!

This commit doesn't actually *implement* any of the new semantics for
different kinds of key lookups, it just exposes them at the reader and
view level, and updates all the callsites so that things compile.