Implement non-blocking range queries in the reader
Support the case where the reader receives a non-blocking, non-equality
KeyComparison, by calling get_range_and on the handle and using serde's
streaming serializer to serialize the results to a buffer.

There are two main more things to implement here - misses, which
requires integrating the interval tree and supporting upqueries, and
blocking reads.